Moved ClaimbuildApplicationRepository to correct folder
