+++ src/main/java/com/ardaslegends/repository/applications/claimbuildapp/ClaimbuildApplicationRepository.java
@@ -1,4 +1,4 @@
-package com.ardaslegends.repository.applications;
+package com.ardaslegends.repository.applications.claimbuildapp;
 
 import com.ardaslegends.domain.applications.ApplicationState;
 import com.ardaslegends.domain.applications.ClaimbuildApplication;
